requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.3.4
asn1crypto==1.4.0
async-timeout==4.0.1
asyncpg==0.24.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
databases==0.5.3
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.70.0
fastapi-mail==1.0.2
fastapi-users==8.1.2
fastapi-users-db-sqlalchemy==1.0.0
greenlet==1.1.2
gunicorn
httpcore==0.13.7
httpx==0.19.0
idna==3.3
Jinja2==3.0.3
makefun==1.12.1
MarkupSafe==2.0.1
packaging==21.2
passlib==1.7.4
pg8000==1.23.0
psycopg2==2.9.2
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==2.4.7
python-multipart==0.0.5
PyYAML==6.0
redis==3.5.3
rfc3986==1.5.0
scramp==1.4.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.26
starlette==0.16.0
uvicorn